+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Rename `teamRating` to `rateTeam` in `Game.Halma.AI.Base`
 - Introduce new type `Move` representing moves on the board
 - changed call signature of `movePiece`
+- nicer default colours
 
 ## 0.2.0.1 (2015-07-22)